+++ tuiml/_cpp/module.cpp
@@ -291,7 +291,7 @@ PYBIND11_MODULE(_cpp_ext, m) {
     m.def("set_num_threads", &tuiml::set_num_threads,
           py::arg("n"), "Set number of OpenMP threads.");
 
-    m.attr("__version__") = "0.1.7";
+    m.attr("__version__") = "0.1.8";
 
 #ifdef TUIML_USE_OPENMP
     m.attr("has_openmp") = true;